requirements.txt
# KPI Platform — Python dependencies
# Install: pip install -r requirements.txt

# Google Sheets API
google-auth>=2.27.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.116.0

# Claude AI API
anthropic>=0.40.0

# Excel report generation
openpyxl>=3.1.2

# HTTP requests (future Zenoti/Salon Ultimate connectors)
requests>=2.31.0
